Finish function implementation and other minor fixes.
Finish slist_move_el implementation. It now passes the tests;
Change numbers added to insertion tests to be more sequential and help to find logic errors;
Fix function name on tests because change of function workflow;
Comment sleep on print_list function for the sake of time execution. It's a test after all;